Refactored rating related sif
