+++ app/build.gradle
@@ -28,7 +28,6 @@ android {
     }
     signingConfigs {
         release {
-            storeFile file(keystorePropertiesFile)
             if (keystorePropertiesFile.exists()) {
                 keyAlias = keystoreProperties["keyAlias"]
                 keyPassword = keystoreProperties["keyPassword"]
